% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_rem.R
\name{attr_rem}
\alias{attr_rem}
\title{Remove attributes}
\usage{
attr_rem(x, which)
}
\arguments{
\item{x}{The object to remove attribute from}

\item{which}{name of the attribute to remove}
}
\value{
"x" without attribute "which"
}
\description{
Remove attributes from objects
}
\details{
This function comes from the package massageR from Jan Stanstrup
}
\references{
https://github.com/stanstrup/massageR
}
\author{
Jan Stanstrup

Rico Derks
}
