% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Create bubble plots}
\usage{
bubble_plot(data, x, y, color_by, carbon_group)
}
\arguments{
\item{data}{the dataframe containing the lipids, normally imported results from MS-DIAL}

\item{x}{what do you want on the x-axis, often the retetnion time}

\item{y}{what do you want on the y-axis, often the retetnion time}

\item{color_by}{color the bubbles according to this column}

\item{carbon_group}{group the bubbles according to this column}
}
\value{
A ggplot2 object
}
\description{
Create a bubble for quick checking of MSMS lipid identification from MS-DIAL
}
\details{
This function will create bubble plots from exported MS-DIAL results.
It will quickly show you if there any misalignment, bad identifications. This function is still heavily
under development and for sure will change in the future. This function assumes the ggplot2 > v2.2.1.9000 is 
installed.
}
\author{
Rico Derks
}
