% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{calc_fa_composition}
\alias{calc_fa_composition}
\title{Calculations for the fatty acid tail composition comparisson}
\usage{
calc_fa_composition(
  lipid_data = NULL,
  meta_data = NULL,
  group_column = NULL,
  selected_group = NULL,
  selected_lipidclass = NULL
)
}
\arguments{
\item{lipid_data}{data.frame containing all the data. Rows are samples, 
columns are variables (lipids). The column names are the names of the 
lipid species in the format of the Sciex Lipidyzer result file.}

\item{meta_data}{data.frame containing all the meta data of the samples.}

\item{group_column}{character(1), column name of the meta data containing the 
group information.}

\item{selected_group}{character(1) vector with the name of one the groups 
used in the comparison.}

\item{selected_lipidclass}{character(1) containing the name of the lipid 
class.}
}
\value{
data.frame in long format.
}
\description{
Calculations to compare the fatty acid tail composition of a lipid class 
between two groups.
}
\author{
Rico Derks
}
