% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_plot.R
\name{diff_plot}
\alias{diff_plot}
\title{Create diff plots}
\usage{
diff_plot(data, x, y, fill_by, facet_y, facet_x)
}
\arguments{
\item{data}{the dataframe containing the lipids, normally imported results from MS-DIAL}

\item{x}{what do you want on the x-axis, often the difference}

\item{y}{what do you want on the y-axis, often the lipids}

\item{fill_by}{factor showing up or down regulation}

\item{facet_y}{facet by this column in y}

\item{facet_x}{facet by this column in x}
}
\value{
A ggplot2 object
}
\description{
Create a plot to show the difference between two samples.
}
\details{
This function will create a difference plot showing up or down regulation.
}
\author{
Rico Derks
}
