% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/equate.R
\name{equate}
\alias{equate}
\title{Quantile equating}
\usage{
equate(data, sequence, test = FALSE)
}
\arguments{
\item{data}{is a matrix containing all the data, rows correspond to protein, columns correspond to samples}

\item{sequence}{is a vector containing the sequence number of each sample}

\item{test}{is a boolean (default false) used to run a test if the equating was succesful. No real equating is done, but the original matrix is exported again.}
}
\value{
a matrix with the equated data
}
\description{
Equating is based on the function *normalize.quantiles* from the library preprocessCore from BioConductor.
}
\author{
Rico Derks
}
\references{
Draisma, H. H. M. et al. Equating, or Correction for Between-Block Effects with Application to Body Fluid LC-MS and NMR Metabolomics Data Sets. Anal. Chem. 82, 1039-1046 (2010).
}

