% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{extract_feature}
\alias{extract_feature}
\title{Extract carbon number and double bond information from lipid species}
\usage{
extract_feature(data = NULL)
}
\arguments{
\item{data}{character vector with lipid species names.}
}
\value{
data.frame containing lipid, lipid class, carbon number of tail 1, 
    carbon number of tail 2, double bond of tail 1, double bond of tail 2,
    total carbon number, total double bond.
}
\description{
Extract carbon number and ouble bond information from lipid species name in 
Sciex Lipidyzer format.
}
\author{
Rico Derks
}
