% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_analysis.R
\name{fa_analysis_calc}
\alias{fa_analysis_calc}
\title{Calculations for the fatty acid analysis plot}
\usage{
fa_analysis_calc(
  lipid_data = NULL,
  meta_data = NULL,
  selected_lipidclass = c("CE", "Cer", "DG", "FA", "HexCER", "LPC", "LPE", "LacCER",
    "PA", "PC", "PE", "PG", "PI", "PS", "SM", "TG", "All", "All_noTG")
)
}
\arguments{
\item{lipid_data}{data.frame}

\item{meta_data}{data.frame, with meta data information.}

\item{selected_lipidclass}{character vector, name of the lipid class(es) of
interest.}
}
\value{
data.frame in wide format.
}
\description{
Calculations for the fatty acid analysis plot.
}
\author{
Rico Derks
}
