% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_analysis.R
\name{fa_analysis_calc_rev}
\alias{fa_analysis_calc_rev}
\title{Calculations for the fatty acid analysis plot}
\usage{
fa_analysis_calc_rev(lipid_data = NULL, meta_data = NULL, selected_fa = NULL)
}
\arguments{
\item{lipid_data}{data.frame}

\item{meta_data}{data.frame, with meta data information.}

\item{selected_fa}{character vector, name of the fatty acid tails of
interest.}
}
\value{
data.frame in wide format.
}
\description{
Calculations for the fatty acid analysis plot.
}
\author{
Rico Derks
}
