% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_analysis.R
\name{fa_analysis_plot}
\alias{fa_analysis_plot}
\title{Fatty acid analysis plot}
\usage{
fa_analysis_plot(
  lipid_data = NULL,
  meta_data = NULL,
  selected_lipidclass = NULL,
  group_column = NULL,
  plot_title = NULL,
  y_axis_title = NULL,
  legend_title = NULL
)
}
\arguments{
\item{lipid_data}{data.frame}

\item{meta_data}{data.frame, with meta data information.}

\item{selected_lipidclass}{character vector, name of the lipid class(es) of
interest.}

\item{group_column}{character(1), column name of the meta data containing the 
group information.}

\item{plot_title}{character(1), title for the plot.}

\item{y_axis_title}{character(1), title for the y-axis.}

\item{legend_title}{character(1), title for the legend.}
}
\value{
Fatty acid analysis plot as ggplot2 object.
}
\description{
Fatty acid analysis plot. Show the amount of fatty acid tail per group.
}
\author{
Rico Derks
}
