% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_analysis.R
\name{fa_analysis_prep}
\alias{fa_analysis_prep}
\title{Prepare the data for the fatty acid analysis plot}
\usage{
fa_analysis_prep(fa_data = NULL, meta_data = NULL, group_column = NULL)
}
\arguments{
\item{fa_data}{data.frame.}

\item{meta_data}{data.frame, with meta data information.}

\item{group_column}{character(1), column name of the meta data containing the 
group information.}
}
\value{
data.frame in long format.
}
\description{
Prepare the data for the fatty acid analysis plot, i.e. transform to long 
format.
}
\author{
Rico Derks
}
