% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{fa_comp_heatmap}
\alias{fa_comp_heatmap}
\title{Heatmap part of the fatty acid composition plot}
\usage{
fa_comp_heatmap(
  hm_data = NULL,
  avg_double = NULL,
  avg_carbon = NULL,
  color_limits = NULL,
  y_pos_right = FALSE,
  show_legend = FALSE
)
}
\arguments{
\item{hm_data}{data.frame with the data.}

\item{avg_double}{numeric(1), average of the total unsaturation.}

\item{avg_carbon}{numeric(1), average of the total carbon number.}

\item{color_limits}{numeric(2), minimum and maxium value for the colour bar.}

\item{y_pos_right}{logical(1), if TRUE the y-axis will be placed on the right
side.}

\item{show_legend}{logical(1), show the legend}
}
\value{
Heatmap as ggplot2 object.
}
\description{
Heatmap part of the fatty acid composition plot.
}
\author{
Rico Derks
}
