% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{fa_histogram}
\alias{fa_histogram}
\title{Create the side histograms}
\usage{
fa_histogram(
  hist_data = NULL,
  position = c("left", "right", "top"),
  title = NULL,
  y_pos_right = FALSE
)
}
\arguments{
\item{hist_data}{data.frame}

\item{position}{characer(1), position of the histogram. 'right' and 'left' 
is for the side histograms.}

\item{title}{character(1), title of the histogram.}

\item{y_pos_right}{logical(1), if TRUE the y-axis will be placed on the right
side.}
}
\value{
Histogram as ggplot2 object.
}
\description{
Create the side histograms for the fatty acid composition.
}
\author{
Rico Derks
}
