\name{flip_peaklist}
\alias{flip_peaklist}
\title{Transpose a peaklist from XCMS}
\usage{
flip_peaklist(peaklist, xset)
}
\arguments{
\item{peaklist}{peaklist create by XCMS}

\item{xset}{xcmsSet object}
}
\value{
The transposed peaklist as dataframe.
}
\description{
This functions transposes a peaklist generated by XCMS.
}
\details{
This functions transposes a peaklist generated by XCMS so that samples (observations) are rows and variables (features) are columns.
}
\author{
Rico Derks
}

