% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndPred.R
\name{IndPred}
\alias{IndPred}
\title{Individual prediction based on a \code{rcpme} model}
\usage{
IndPred(rcpmeObj)
}
\arguments{
\item{rcpmeObj}{An object which contains the output of a \code{rcpme} function call on which individual predictions are to be computed.}
}
\value{
A list with each individual prediction
}
\description{
Individual prediction based on a \code{rcpme} model
}
