% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_pow_x.R
\name{m_pow_x}
\alias{m_pow_x}
\title{Modeling power}
\usage{
m_pow_x(data, pca_model)
}
\arguments{
\item{data}{is a matrix of the original data (centered and scaled)}

\item{pca_model}{is the pca model from the package pcaMethods}
}
\value{
Vector of the modeling power for each variable
}
\description{
Calculate the modeling power of each variable from a PCA model.
}
\details{
Calculates the modeling power of each variable from a PCA model as is done in Umetrics Simca.
}
\author{
Rico Derks
}
