% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msca_p.R
\name{msca_p}
\alias{msca_p}
\title{Multilevel Simultaneous Component Analysis (MSCA-P)}
\usage{
msca_p(X, persID)
}
\arguments{
\item{X}{a data table with structure ((I x K) x J), see details for more information}

\item{persID}{vector length (IxK) person ids (numeric e.g. for the situation with 4 samples per person 1 1 1 1 2 2 2 2 3 3 3 3, or factor)}
}
\value{
A data frame with the following structure :
\item{$between$data}{table used for the between-individual level}
\item{$between$scores}{between-individual scores}
\item{$between$loadings}{between-individual loadings}
\item{$between$percexp}{percentage of variation explained in the between-individual model}
\item{$within[[i]]$data}{table used for the within-individual level for individual i}
\item{$within[[i]]$scores}{within-individual scores for individual i}
\item{$within[[i]]$loadings}{within-individual loadings for individual i}
\item{$within[[i]]$percexp}{percentage of variation explained in the within-individual model for individual i}
}
\description{
Multilevel Simultaneous Component Analysis (MSCA-P).
}
\details{
Unlike MSCA, MSCA-P does not make any assumptions about the similarities between the time-dynamic variations of different individuals, 
so scores and loadings will be computed separately for each individual. X needs to have the following data table structure ((I x K) x J). 
Where I are the individuals, J are the variables, and K are samples of an individual.

Note: msca_p can NOT handle NA's!

Based on the algorithm Jansen et al. ANALYTICA CHIMICA ACTA (2005) 530(2):173-183
Based on MatLab code available here http://www.bdagroup.nl/content/Downloads/software/software.php
}
\references{
Jansen et al. ANALYTICA CHIMICA ACTA (2005) 530(2):173-183
}
\author{
E. Nevedomskaya
}
