% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ellipse.R
\name{multi_ellipse}
\alias{multi_ellipse}
\title{Create a Hotelling T2 ellipse for a PCA score plot per group}
\usage{
multi_ellipse(data, x, y, alpha = 0.95, len = 200, ellipse = NULL)
}
\arguments{
\item{data}{data frame or table}

\item{x}{(name of) x vector}

\item{y}{(name of) y vector}

\item{alpha}{confidence interval}

\item{len}{number of points to create the ellipse}

\item{ellipse}{name of column of groups}
}
\value{
A data frame is returned with the points to create the ellipse.
}
\description{
This function can be used to create a confidence (Hotelling T2) interval for a 
PCA score plot per group.
}
\details{
This is a helper function which is used to create a confidence (Hotelling T2) interval for a 
PCA score plot per group.
}
\author{
Rico Derks, modified by Marco Bladergroen

# if ellipse is not NULL this function assumes you have data with a factorial column named as in 'ellipse' to draw multiple ellipses.
x and y should contain the name (string) of the columns of the data to be plotted. If ellipse is NULL the (single) ellipse is drawn according to data in x and y (vectors).
}
