% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function (Deprecated)}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{optionally ggplot objects}

\item{plotlist}{list of ggplots}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present, \code{cols} is ignored. See for more info Details}
}
\value{
One big graph containing multiple ggplot graphs.
}
\description{
Put multiple graphs on one page.
}
\details{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects) 
If the layout is something like matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE), then plot 1 will go in the upper left, 2 will go in the upper right, and 
3 will go all the way across the bottom.
}
\references{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\author{
Rico Derks
}
