% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls_pred.R
\name{opls_pred}
\alias{opls_pred}
\title{Prediction function of the opls package}
\usage{
opls_pred(A, model, direction)
}
\arguments{
\item{A}{Observations}

\item{model}{opls model}

\item{direction}{from x to y ('xy') or from y to x ('yx')}
}
\value{
Matrix :
\item{Bhat}{Predictions}
\item{Scorenew}{Corresponding scores}
}
\description{
Prediction function of the opls package.
}
\details{
Prediction function of the opls package.
}
\author{
E. Nevedomskaya

Rico Derks
}
