% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{plot_fa_composition}
\alias{plot_fa_composition}
\title{Compare the fatty acid tail composition}
\usage{
plot_fa_composition(
  lipid_data = NULL,
  meta_data = NULL,
  group_column = NULL,
  groups = NULL,
  selected_lipidclass = NULL
)
}
\arguments{
\item{lipid_data}{data.frame containing all the data.}

\item{meta_data}{data.frame containing all meta data.}

\item{group_column}{character(1) column name (meta data) containing all 
column information.}

\item{groups}{character(2) the name of the two groups for the comparison.}

\item{selected_lipidclass}{character(1) name of the lipid class of interest.}
}
\value{
ggplot2 object.
}
\description{
Compare the fatty acid tail composition of a lipid class between two groups.
}
\author{
Rico Derks
}
