% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_composition.R
\name{prep_histogram}
\alias{prep_histogram}
\title{Prepare data for carbon number histogram}
\usage{
prep_histogram(data = NULL, position = c("side", "top"))
}
\arguments{
\item{data}{data.frame}

\item{position}{character(1), prepare the data for the 'top' or 'side' 
histogram.}
}
\value{
data.frame in long format.
}
\description{
Prepare the data to create the carbon number histogram
}
\author{
Rico Derks
}
