% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Center and scale your data}
\usage{
preprocess(
  data,
  center = TRUE,
  scale = c("none", "uv", "pareto"),
  simple = TRUE
)
}
\arguments{
\item{data}{matrix to preprocess}

\item{center}{center the data (default is TRUE)}

\item{scale}{which scaling to use \code{c("none", "uv", "pareto")}}

\item{simple}{only return the preprocessed matrix. If false also the original 
data will be returned as a list. Default is true.}
}
\value{
The centered and scaled matrix will be returned if simple is true. If simple is false also the center and scale vectors are returned.
}
\description{
With this function you can center and scale your data.
}
\author{
Rico Derks
}
