% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{qc_plot}
\alias{qc_plot}
\title{Create shewart chart like plots}
\usage{
qc_plot(data, x, y, color_by, avg, stdev, rsd, xlabel = "", ylabel = "")
}
\arguments{
\item{data}{the dataframe containing the lipids, normally imported results from MS-DIAL}

\item{x}{what do you want on the x-axis, often the retetnion time}

\item{y}{what do you want on the y-axis, often the retetnion time}

\item{color_by}{color used if you have multiple variables to show, it will also be used for facetting}

\item{avg}{column which contains the average values}

\item{stdev}{column which contains the standard devation values}

\item{rsd}{column which contains the values for the RSD / CV}

\item{xlabel}{define the label for the x-axis}

\item{ylabel}{define the label for the y-axis}
}
\value{
A ggplot2 object
}
\description{
Create Shewart chart like plots
}
\details{
This function will give facetted Shewart chart like plots. With color_by each 
line (variable) will get a different color. This will also used for facetting. 
This function is still heavily under development and for sure will change in the future. 
This function assumes the ggplot2 > v2.2.1.9000 is installed.
}
\author{
Rico Derks
}
