% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpm.R
\docType{package}
\name{Rcpm-package}
\alias{Rcpm}
\alias{Rcpm-package}
\title{General functions for the Center of Proteomics and Metabolomics (LUMC)}
\description{
This package contains some general functions for use within the Center of 
Proteomics and Metabolomics
}
\details{
It contains all kind of functions. Most of them still need examples.

A vignette is written for targeted analysis with xcms.
}
