% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rcpm.R
\docType{package}
\name{Rcpm}
\alias{Rcpm}
\alias{Rcpm-package}
\title{Rcpm: A package with general functions.}
\description{
A package with general functions used within the Center for Proteomics
and Metabolomics
}
\section{Functions}{

\itemize{
  \item \code{\link{equate}}
  \item \code{\link{merge_xcms_data}}
  \item \code{\link{pca_plots}}
  \item \code{\link{pqn}}
  \item \code{\link{preprocess}}
  \item \code{\link{qc_rlsc}}
  \item \code{\link{simple_ellipse}}
  \item \code{\link{summary_se}}
  \item \code{\link{vip}}
  \item \code{\link{vip_jh}}
}
}

