% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpme.R
\name{rcpme}
\alias{rcpme}
\title{Random Change Point Mixed Model}
\usage{
rcpme(longdata, formu, covariate = "NULL", REadjust = "no",
  gamma = 0.1, nbnodes = 10, param = NULL)
}
\arguments{
\item{longdata}{A dataframe containing the variables used in the formula  \code{formu}}

\item{formu}{A formula object describing which variables are to be used. The formula has to be of the following form \code{markervar ~ scorevar | grouvpar} for the function to work.}

\item{covariate}{An optional string indicating a binary covariate to add on the fixed effects, i.e. intercept, mean slope, difference of slopes and changepoint date. The parameter \code{REadjust} indicates how this covariate influences the random effects variance structure. Default to NULL, i.e. no covariates.}

\item{REadjust}{An optional string value indicating how the random effects variance structure depends on \code{covariate}. "no" means that the structure doesn't depend upon \code{covariate}. "prop" indicates that the random effects variance structure is proportionnal according to \code{covariate} value. "yes" indicates that there is two different random effects variance structures, i.e. one for each level of \code{covariate}. Default to "no".}

\item{gamma}{A numeric parameter indicating how smooth the trajectory is on the changepoint date. Default to 0.1.}

\item{nbnodes}{A numeric parameter indicating how many nodes are to be used for the gaussian quadrature for numerical integration. Default to 10.}
}
\value{
The output contains several objects : \code{loglik} is the value of the log-likelihood at the optimum; \code{fixed} contains all fixed parameters estimates, standard errors, CIs, wald test statistic and corresponding pvalue when possible; \code{sdres} the estimated residual error; \code{VarEA} a 4x4 matrix or a list of 4x4 matrices - if there is some covariate for example - containing the estimated random effects covariance matrix; \code{optpar} the optimal parameters maximizing the log-likelihood; \code{covariate} the covariate declared in the function call; \code{REadjust} the string indicating how random effects structure is handled as declared in the function call, \code{invhessian} the covariance matrix containing all the standard errors and correlations of the parameter estimates;
}
\description{
Random Change Point Mixed Model
}
