% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgf.R
\name{read_mgf}
\alias{read_mgf}
\title{Read mgf file}
\usage{
read_mgf(filename = NULL)
}
\arguments{
\item{filename}{filename of the mgf file.}
}
\value{
a dataframe
}
\description{
Read a mgf file.
}
\details{
An extra column **idx** is created in the data frame.
}
\author{
Rico Derks
}
