% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REestimate.R
\name{REestimate}
\alias{REestimate}
\title{Random Effects estimate}
\usage{
REestimate(rcpmeObj, longdata, var = FALSE, onlytau = FALSE)
}
\arguments{
\item{rcpmeObj}{An object which contains the output of a \code{rcpme} function call.}

\item{var}{A boolean indicating if the covariance matrices of random effects estimation should be returned. Default to \code{FALSE}.}
}
\value{
A list with each item containing the subject specific random effects estimation and its covariance matrix if \code{var = TRUE}.
}
\description{
Random Effects estimate
}
