% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fit.R
\name{summary_fit}
\alias{summary_fit}
\title{show summary of fit of a PCA model}
\usage{
summary_fit(data)
}
\arguments{
\item{data}{data frame with all data. See details for more information.}
}
\value{
A ggplot2 plot is returned.
}
\description{
Create a plot which shows a summary of the fit your PCA model.
}
\details{
The data frame \code{data} should contain 3 variables. \code{PC}, 
\code{R2cum} and \code{Q2cum}.
}
\examples{
library(pcaMethods)

# create some dummy data
my_data <- data.frame(samples = c(rep("a", 10), rep("b", 10)),
                      matrix(data = rnorm(400),
                             nrow = 20))

# as cross validation q2 needs to be selected
M1 <- pca(object = my_data,
          nPcs = 5,
          cv = "q2")
          
# create data frame with R2 and Q2            
sumfit_data <- data.frame(PC = paste("PC", 1:5, sep = ""), 
                          R2cum = M1@R2cum, 
                          Q2cum = M1@cvstat)

# create the summary plot
p <- summary_fit(sumfit_data)
}
\author{
Rico Derks
}
