% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2roi.R
\name{tbl2roi}
\alias{tbl2roi}
\title{Convert a list of peaks (rt / m/z pairs) to a Region of Interest (ROI) list for use
with \code{\link{findPeaks}}}
\usage{
tbl2roi(tbl, raw, mz_tol = 0.005, mz_tol_unit = c("Da", "ppm"),
  rt_tol = 10)
}
\arguments{
\item{tbl}{\code{\link{tibble}} containing the columns "rt" and "mz". rt needs to be in seconds.}

\item{raw}{\code{\link{xcmsRaw}} object to create ROI for. It needs to be a specific 
\code{\link{xcmsRaw}} to match retention times to scan nubmers.}

\item{mz_tol}{m/z tolorance}

\item{mz_tol_unit}{ppm tolerance for the generated ROI.}

\item{rt_tol}{Retention time tolerance (in seconds!) for the generated ROI.}
}
\value{
List containing the ROIs. Each list contains m/z, mzmin, mzmax, scmin, scmax, 
length (set to -1, not used by centWave) and intensity (set to -1, not used by centWave) columns.
}
\description{
Convert a list of peaks (rt / m/z pairs) to a Region of Interest (ROI) list for use
with \code{\link{findPeaks}}
}
\details{
This is based on the work of Jan Stanstrup (see references).
}
\references{
https://github.com/stanstrup/QC4Metabolomics
}
\author{
Jan Stanstrup

Rico Derks
}
