% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRCPMM.R
\name{testRCPMM}
\alias{testRCPMM}
\title{testRCPMM}
\usage{
testRCPMM(longdata, formu, covariate = NULL, gamma = 0.1,
  nbnodes = 5, nbpert = 500)
}
\arguments{
\item{longdata}{Longitudinal data}

\item{formu}{A formula indicating which variables from longdata should be used. It must have the following form : markervar ~ timevar | groupvar}

\item{gamma}{A smoothing parameter for the transition on the changepoint date. 0.1 by default.}

\item{nbnodes}{Number of adaptive quadrature nodes used to compute the integrals. 5 by default.}

\item{nbpert}{Number of perturbations used to compute the empirical p-value. 500 by default.}
}
\value{
The function returns a list with the computed empirical p-value, the nbpert perturbed test statistics and the observed test statistic.
}
\description{
Realizes the supremum score test on longdata according to formu.
}
