% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf.R
\name{write_mgf}
\alias{write_mgf}
\title{Write mgf file}
\usage{
write_mgf(filename = NULL, mgf_data = NULL)
}
\arguments{
\item{filename}{filename of the mgf file}

\item{mgf_data}{data frame containing all the data. Each row is a MSMS spectrum.}
}
\description{
Write a dataframe to a mgf file.
}
\details{
Columns which will be written to the mgf file: scans, pepmass, mslevel, 
    charge, rtminutes, ion, msms_spectrum
}
\author{
Rico Derks
}
