% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcmsraw_to_tbl.R
\name{xcmsraw_to_tbl}
\alias{xcmsraw_to_tbl}
\title{Convert raw data into a tibble of xcmsRaw objects.}
\usage{
xcmsraw_to_tbl(files, ...)
}
\arguments{
\item{files}{character vector of file names/paths.}

\item{...}{further arguments to \code{\link{xcmsRaw}}.}
}
\value{
A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{sample_name:} Filename without path.
  \item \strong{raw:} The xcmsRaw objects.
}
}
\description{
Convert raw data into a tibble of xcmsRaw objects.
}
\details{
This will read multiple files with xcmsRaw and put all xcmsRaw objects in a table.
}
\author{
Rico Derks
}
