% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Create bubble plots}
\usage{
bubble_plot(data, x, y, color_by, carbon_db)
}
\arguments{
\item{data}{Dataframe containing all information. See details for more information on the structure.}

\item{x}{what to plot on the x-axis, in general retention time.}

\item{y}{what to plot on the y-axis, in general m/z.}

\item{color_by}{color the lines, in general the number carbons.}

\item{carbon_db}{column containing the total carbon numer and total number of double bounds}
}
\value{
A ggplot2 plot is returned.
}
\description{
Create bubble plots from MS-DIAL lipidomics result files.
}
\details{
\code{data} should be data frame which contains at least all the parameters as columns.
}
\examples{
my_data <- data.frame(rt = rep(c(seq(7, 10, length.out = 5),
                                 seq(3, 6, length.out = 3), 
                                 seq(2, 5, length.out = 7),
                                 seq(2, 8, length.out = 6),
                                 seq(5, 9, length.out = 4),
                                 seq(3, 10, length.out = 6)), 3),
                      mz = rep(seq(507, 510, by = 0.1), 3),
                      carbon = c(rep("43", 5), 
                                 rep("41", 3), 
                                 rep("44", 7), 
                                 rep("46", 6), 
                                 rep("42", 4), 
                                 rep("45", 6)),
                      carbon_db = c(paste0(rep("43", 5), ":", 1:5),
                      paste0(rep("41", 3), ":", 1:3),
                      paste0(rep("44", 7), ":", 1:7),
                      paste0(rep("46", 6), ":", 1:6),
                      paste0(rep("42", 4), ":", 1:4),
                      paste0(rep("45", 6), ":", 1:6)),
                      DotProduct = rnorm(n = 93, mean = 700, sd = 200),
                      LipidClass = c(rep("a", 31), rep("b", 31), rep("c", 31)))
# create the bubble plot
bubble_plot(data = my_data,
            x = rt,
            y = mz,
            color_by = carbon,
            carbon_db = carbon_db)
}
\author{
Rico Derks
}
