% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment.R
\name{check_alignment}
\alias{check_alignment}
\title{Check the alignment of mzXML files}
\usage{
check_alignment(files)
}
\arguments{
\item{files}{A character vector containing all the mzXML file locations.}
}
\value{
The function gives a message with the file name of the problematic mzXML file and returns TRUE or FALSE.
}
\description{
Check the aligned mzXML files for problematic retention times.
}
\details{
msalign2 sometimes creates wrong retention times. The retention time of a scan is sometimes smaller then its previous scan.
}
\author{
Rico Derks
}
