% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_plot.R
\name{diff_plot}
\alias{diff_plot}
\title{Create difference plots}
\usage{
diff_plot(data, x, y, fill_by, facet_y, facet_x)
}
\arguments{
\item{data}{Dataframe containing all information. See details for more information on the structure.}

\item{x}{what to plot on the x-axis, in general retention time.}

\item{y}{what to plot on the y-axis, in general m/z.}

\item{fill_by}{column with factor for filling the bar plot}

\item{facet_y}{facet in y direction with column}

\item{facet_x}{facet in x direction with column}
}
\value{
A ggplot2 plot is returned.
}
\description{
Create difference plots to show up- or down-regulation of certain lipids.
}
\details{
\code{data} should be data frame which contains at least all the parameters as columns.
}
\examples{
# create the diff plot
diff_plot(data = my_data,
            x = LongLipidName,
            y = difference,
            fill_by = diff_grp,
            facet_y = LipidClass,
            facet_x = versus)
}
\author{
Rico Derks
}
