% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mzxml_spectra.R
\name{get_mzxml_spectra}
\alias{get_mzxml_spectra}
\title{Retrieve MS1 spectra from a mzXML file}
\usage{
get_mzxml_spectra(file)
}
\arguments{
\item{file}{mzXML file}
}
\value{
The function returns a list of spectra. A spectrum is a matrix where the first column is \emph{m/z} and the second column is intensity.
}
\description{
Retrieve all MS1 spectra from a mzXML file.
}
\details{
This is first version for retrieving spectra from mzXML files. For now it only works on MS1 data, there is no error checking. It assumes that the 
presicion is 64bit, endian is network and the compression is zlib.
}
\author{
Rico Derks
}
