% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msms_search_lipidmaps.R
\name{msms_search_lipidmaps}
\alias{msms_search_lipidmaps}
\title{Do a MSMS search on the Glycerophospholipid MS/MS Prediction database from LipidMaps}
\usage{
msms_search_lipidmaps(
  lipidclass = c("glycerolipids", "phospholipids"),
  intensity_threshold = 10,
  prec_tol = 0.01,
  prod_tol = 0.01,
  headgroup = "",
  ion = "",
  min_matches = 5,
  LIMIT = c("All", "even"),
  peaklist = NULL
)
}
\arguments{
\item{lipidclass}{define the lipidclass you want to search, currently available phospholipics and glycerolipids}

\item{intensity_threshold}{set the intensity threshold}

\item{prec_tol}{precursor ion mass tolerance (+/- m/z)}

\item{prod_tol}{product ion mass tolerance (+/- m/z)}

\item{headgroup}{search all headgroups of a lipidclass, see details for more inforamtion}

\item{ion}{set the ion type (see details for more info)}

\item{min_matches}{how many matches should the search at least have}

\item{LIMIT}{limit the chains to 'even' or 'odd' and 'even' (All)}

\item{peaklist}{in PKL format
See detail section for more information.}
}
\value{
The function returns a dataframe with the search results. If there is a connection problem NULL is returned and a 
warning message showing the status code of the connection.
}
\description{
Search the Glycerophospholipid MS/MS Prediction database from LipidMaps (http://www.lipidmaps.org/).
}
\details{
This function uses the Programmatic Access functionality of LipidMaps to search the LMSD database.

Phospholipids:

Headgroups are defined as:  "Any" or one off "PC", "PA", "PS", "PE", "PG", "PI"

The ion types are defined as: 
* chloride : [M-H]- / [M+Cl]-
* acetate : [M-H]- / [M+Acetate-H]-
* formate : [M-H]- / [M+Formate-H]-

Glycerolipids:

Headgroups are defined as:  "Any" or one off "MD", "DG", "TG"

The ion types are defined as: 
* ammonium : [M+NH4]+
}
\examples{
mypeaks <- c("818.59 10000 1\n758.57 20\n687.50 20\n506.33 50\n494.33 15\n488.31 2")

results <- msms_search_lipidmaps(lipidclass = "phospholipids",
                     intensity_threshold = 10,
                     prec_tol = 0.01,
                     prod_tol = 0.01,
                     headgroup = "Any",
                     ion = "formate",
                     min_matches = 2,
                     LIMIT = "All",
                     peaklist = mypeaks)
}
\author{
Rico Derks
}
