\name{o2pls_pred}
\alias{o2pls_pred}
\title{Prediction function of the opls package}
\usage{
o2pls_pred(A, model, direction)
}
\arguments{
\item{A}{Observations}
\item{model}{o2pls model}
\item{direction}{from x to y ('xy') or from y to x ('yx')}
}
\description{
Prediction function of the opls package
}
\details{
Prediction function of the opls package
}
\value{
\item{Bhat}{Predictions}
\item{Scorenew}{Corresponding scores}
}
\author{
E. Nevedomskaya, Rico Derks
}