% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{pca_plots}
\alias{pca_plots}
\title{create PCA plots}
\usage{
pca_plots(data, xPC = 1, yPC = 2, type = c("scores", "loadings",
  "summary"), T2 = TRUE, hotelling = 0.95, show.groups = FALSE,
  group.names, group.ids, ...)
}
\arguments{
\item{data}{Dataframe containing all information. See details for more
information on the structure.}

\item{xPC}{Principal component for the x-axis.}

\item{yPC}{Principal component for the y-axis.}

\item{type}{Which plot to generate: scores, loadings or summary plot.}

\item{T2}{add a Hotteling T2 ellipse to the scores plot. Default is true.}

\item{hotelling}{Set the Hotelling confidence interval. Default = 0.95.}
}
\description{
Generate different plots for a PCA analysis. Score plot
  (includings Hotelling T2 confidence interval ellipse), loading plot or a
  summary plot. The summary plot contains a score plot, loadings plot and the
  summary of fit plot.
}
\author{
Rico Derks
}

