% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot nice histogram}
\usage{
plot_histogram(data, variable = NULL, legend = TRUE)
}
\arguments{
\item{data}{is a data frame containing all the data}

\item{variable}{which variables (i.e. columns) to plot (see examples).}

\item{legend}{is a boolean (default true) used to show the legend.}
}
\value{
ggplot showing a nice histogram
}
\description{
plot_histogram plots a histogram with a fitted kernel density plot and a normal 
density plot. If no \code{variable} (i.e. NULL) is defined, then for each column in the 
data frame a histogram is plotted.
}
\examples{
set.seed(123)
my_df <- data.frame(a = rnorm(100, mean = 10), 
                    b = rnorm(100, mean = 40),
                    c = rnorm(1000, mean = 123))
                    
# all variables
plot_histogram(data = my_df)

# only one
plot_histogram(data = my_df,
               variable = a)
               
# only two
plot_histogram(data = my_df,
               variable = c(a, b))

# or using tidyverse approach
library(dplyr)
my_df \%>\%
  plot_histogram()
  
}
\author{
Rico Derks
}
