% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pqn.R
\name{pqn}
\alias{pqn}
\title{Perform Probabilistic Quotient Normalization (Median Fold Change)}
\usage{
pqn(X, n = "median", QC)
}
\arguments{
\item{X}{matrix to normalize samples * variables (rows * columns)}

\item{n}{normalization reference: "mean" for using the overall average of variables as reference or "median" (default) for using the overall median of variables as reference}

\item{QC}{vector of number(s) to specify samples which average to use as reference (e.g. QC samples)}
}
\value{
Normalized table samples * variables (rows * columns)
}
\details{
First a total area normalization should be done before PQN is applied.
}
\description{
Perform Probabilistic Quotient Normalization (Median Fold Change)
}
\author{
E. Nevedomskaya
}
\references{
1.Dieterle, F., Ross, A., Schlotterbeck, G. & Senn, H. Probabilistic Quotient Normalization as Robust Method to Account for Dilution of Complex Biological Mixtures. Application in H1 NMR Metabonomics. Anal. Chem. 78, 4281-4290 (2006).
}

