% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Center and scale your data}
\usage{
preprocess(data, center = TRUE, scale = c("none", "uv", "pareto"),
  reverse = FALSE, simple = TRUE)
}
\arguments{
\item{data}{matrix to preprocess}

\item{center}{center the data (default is TRUE)}

\item{scale}{which scaling to use \code{none}, \code{uv} or \code{pareto}}

\item{reverse}{reverse the scaling \\ centering}

\item{simple}{only return the preprocessed matrix. If false also the original data will be returned as a list.}
}
\value{
The centered and scaled matrix will be returned.
}
\description{
Center and scale your data
}

