% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{qc_plot}
\alias{qc_plot}
\title{Create QC plots}
\usage{
qc_plot(data, x, y, color_by, avg, stdev, rsd, xlabel = "", ylabel = "")
}
\arguments{
\item{data}{Dataframe containing all information. See details for more information on the structure.}

\item{x}{what to plot on the x-axis, in general something with a time component in it.}

\item{y}{what to plot on the y-axis, in general something like peak area or signal height.}

\item{color_by}{color the lines and facet according to this column.}

\item{avg}{the column with the average per compound.}

\item{stdev}{the column with the standard deviation per compound.}

\item{rsd}{the column with the RSD / CV per compound, this will be used in the title of the facets.}

\item{xlabel}{set the label of the x-axis.}

\item{ylabel}{set the label of the y-axis.}
}
\value{
A ggplot2 plot is returned.
}
\description{
Create Shewhary Chart like QC plots for our metabolomics workflow.
}
\details{
\code{data} should be data frame which contains at least all the parameters as columns.
}
\examples{
require(dplyr)
set.seed(11)
# create a dummy frame 
my_data <- data.frame(x = rep(1:5, 8),
                      y = rnorm(40, mean = 10),
                      group = as.factor(rep(1:5, 8)))
# calculate average / sd / rsd                       
my_data <- my_data \%>\%
  group_by(group) \%>\%
  mutate(avg = mean(y),
         stdev = sd(y),
         rsd = stdev / avg * 100)
# make the QC plot          
p <- qc_plot(data = my_data,
             x = x,
             y = y,
             color_by = group,
             avg = avg,
             stdev = stdev,
             rsd = rsd,
             xlabel = "Measurement index",
             ylabel = "peak area")
}
\author{
Rico Derks
}
