% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rlsc.R
\name{qc_rlsc}
\alias{qc_rlsc}
\title{Perform QC-RLSC for batch correction of chromatographic signal}
\usage{
qc_rlsc(tab, colv, or, verbose = FALSE)
}
\arguments{
\item{tab}{table N*K (row * column) with N samples and K variables.}

\item{colv}{vector N of numbers: 1 for QC samples and 2 for other samples.}

\item{or}{vector of measuring order (see details).}

\item{verbose}{print which variable has been corrected to monitor the process (default = FALSE).}
}
\value{
corrected table N*K
}
\description{
Perform QC-RLSC for batch correction of chromatographic signal.
}
\details{
Make sure that everything is sorted in measurement order!!!
}
\references{
Dunn et al. Nature Protocols 6, 1060-1083 (2011)
}
\author{
E. Nevedomskaya

Rico Derks
}
