\name{rcpm-package}
\alias{rcpm-package}
\alias{rcpm}
\docType{package}
\title{
\packageTitle{rcpm}
}
\description{
\packageDescription{rcpm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcpm}
\packageIndices{rcpm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rcpm}

Maintainer: \packageMaintainer{rcpm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
