% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpm.R
\docType{package}
\name{Rcpm}
\alias{Rcpm}
\alias{Rcpm-package}
\title{Rcpm: A package with general functions.}
\description{
A package with general functions used within the Center for Proteomics
and Metabolomics
}
\section{Functions}{

\itemize{
		\item \code{\link{check_alignment}} Check the aligned mzXML files for problematic retention times
	\item \code{\link{equate}} quantile equate
		\item \code{\link{flip_peaklist}} This functions transposes a peaklist generated by XCMS so that samples (observations) are rows and variables (features) are columns.
	\item \code{\link{get_mzxml_spectra}} Retrieve all MS1 spectra from a mzXML file
		\item \code{\link{m_pow_x}} Modeling power
	\item \code{\link{merge_xcms_meta}} merge xcms output file with meta data
	\item \code{\link{msca}} multilevel simultaneous analysis
	\item \code{\link{ms_search_lipidmaps}} MS search on the LMSD database from LipidMaps
	\item \code{\link{ncomp_opls}} Find the optimal number of OPLS components (only works for single Y)
	\item \code{\link{o2pls_pred}} Prediction function of the opls package
	\item \code{\link{o2pls}} Orthogonal Projections to Latent Structures (OPLS) for a multiple Y's
	\item \code{\link{opls_pred}} Prediction function of the opls package
	\item \code{\link{opls}} Orthogonal Projections to Latent Structures (OPLS) for a single Y
	\item \code{\link{pca_plots}} plot several PCA plots
	\item \code{\link{perm_test}} Permutation test for PLS and PLS-DA models
	\item \code{\link{plot_histogram}} plot nice histogram
	\item \code{\link{pqn}} Perform Probabilistic Quotient Normalization
	\item \code{\link{preprocess}} Centering / scaling
	\item \code{\link{qc_rlsc}} LOESS correction
	\item \code{\link{simple_ellipse}} Hotelling T2 ellipse
	\item \code{\link{summary_fit}} Plot a summary of fit of a PCA model
	\item \code{\link{summary_fit_data}} Restructure the data for a summary of fit plot
	\item \code{\link{summary_se}}
	\item \code{\link{vip}}
	\item \code{\link{vip_jh}}
	\item \code{\link{xcmsraw_to_tbl}} Convert raw data into a tibble of xcmsRaw objects.
}
}

