\docType{package}
\name{Rcpm}
\alias{Rcpm}
\alias{Rcpm-package}
\title{Rcpm: A package with general functions.}
\description{
A package with general functions used within the Center for Proteomics
and Metabolomics
}
\section{Functions}{

\itemize{
  \item \code{\link{equate}} quantile equate
  \item \code{\link{merge_xcms_meta}} merge xcms output file with meta data
  \item \code{\link{msca}} multilevel simultaneous analysis
  \item \code{\link{ncomp_opls}} Find the optimal number of OPLS components (only works for single Y)
  \item \code{\link{o2pls_pred}} Prediction function of the opls package
  \item \code{\link{o2pls}} Orthogonal Projections to Latent Structures (OPLS) for a multiple Y's
  \item \code{\link{opls_pred}} Prediction function of the opls package
  \item \code{\link{opls}} Orthogonal Projections to Latent Structures (OPLS) for a single Y
  \item \code{\link{pca_plots}} plot several PCA plots
  \item \code{\link{perm_test}} Permutation test for PLS and PLS-DA models
  \item \code{\link{pqn}} Perform Probabilistic Quotient Normalization
  \item \code{\link{preprocess}} Centering / scaling
  \item \code{\link{qc_rlsc}} LOESS correction
  \item \code{\link{simple_ellipse}} Hotelling T2 ellipse
  \item \code{\link{summary_fit}} Plot a summary of fit of a PCA model
  \item \code{\link{summary_fit_data}} Restructure the data for a summary of fit plot
  \item \code{\link{summary_se}}
  \item \code{\link{vip}}
  \item \code{\link{vip_jh}}
}
}

